Added a check for the existance of a settings gradle file. fixes #54

+++ lib/dryrun/android_project.rb
@@ -77,6 +77,7 @@ module Dryrun
 
     def find_modules
       return [] unless is_valid
+      return [] unless File.exist?(@settings_gradle_path)
 
       content = File.open(@settings_gradle_path, "rb").read
       modules = content.scan(/'([^']*)'/)
